# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Meridian API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MoveMoneyInput {
  /** Approximate amount to move */
  amountRange: string;
  /** Current location of funds */
  fundsLocation: string;
  /** Intended destination for funds */
  destination: string;
  /** Purpose of the transfer */
  purpose: string;
  /** When looking to proceed */
  timeline: string;
  /** Whether funds held with regulated institution */
  regulatedInstitution: string;
  fullName: string;
  countryOfResidence: string;
  whatsappNumber: string;
  emailAddress: string;
}

export interface PrivateBankingInput {
  /** Type of banking relationship sought */
  bankingRelationshipType: string;
  /** Approximate investable wealth */
  investableWealth: string;
  /** Jurisdictions of interest (multiple) */
  jurisdictions: string[];
  /** Main reasons for seeking private banking (multiple) */
  reasons: string[];
  /** How soon to establish relationship */
  timeline: string;
  fullName: string;
  countryOfResidence: string;
  whatsappNumber: string;
  emailAddress: string;
}

export interface EnquiryResult {
  /** Generated reference number for the enquiry */
  referenceNumber: string;
  /** Client full name */
  fullName: string;
  /** Enquiry type: move-money or private-banking */
  type: string;
  /**
     * WhatsApp number for the desk
     * @nullable
     */
  whatsappNumber?: string | null;
  /**
     * wa.me link for the desk
     * @nullable
     */
  whatsappLink?: string | null;
}

export interface ErrorResponse {
  error: string;
}