# lib/api-zod/src/generated/types/enquiryResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Meridian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryResult {
  /** Generated reference number for the enquiry */
  referenceNumber: string;
  /** Client full name */
  fullName: string;
  /** Enquiry type: move-money or private-banking */
  type: string;
  /**
     * WhatsApp number for the desk
     * @nullable
     */
  whatsappNumber?: string | null;
  /**
     * wa.me link for the desk
     * @nullable
     */
  whatsappLink?: string | null;
}