# lib/api-zod/src/generated/types/moveMoneyInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Meridian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MoveMoneyInput {
  /** Approximate amount to move */
  amountRange: string;
  /** Current location of funds */
  fundsLocation: string;
  /** Intended destination for funds */
  destination: string;
  /** Purpose of the transfer */
  purpose: string;
  /** When looking to proceed */
  timeline: string;
  /** Whether funds held with regulated institution */
  regulatedInstitution: string;
  fullName: string;
  countryOfResidence: string;
  whatsappNumber: string;
  emailAddress: string;
}