# lib/api-zod/src/generated/types/privateBankingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Meridian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PrivateBankingInput {
  /** Type of banking relationship sought */
  bankingRelationshipType: string;
  /** Approximate investable wealth */
  investableWealth: string;
  /** Jurisdictions of interest (multiple) */
  jurisdictions: string[];
  /** Main reasons for seeking private banking (multiple) */
  reasons: string[];
  /** How soon to establish relationship */
  timeline: string;
  fullName: string;
  countryOfResidence: string;
  whatsappNumber: string;
  emailAddress: string;
}